chore: forge fmt

+++ src/test/mocks/MockBN254CertificateVerifier.sol
@@ -112,17 +112,15 @@ contract MockBN254CertificateVerifier is IBN254CertificateVerifier {
         return true;
     }
 
-    function getTotalStakeWeights(
-        OperatorSet memory, /*operatorSet*/
-        uint32 /*referenceTimestamp*/
-    ) external pure returns (uint256[] memory) {
-        return new uint256[](0);
+    function getTotalStakeWeights(OperatorSet memory, /*operatorSet*/ uint32 /*referenceTimestamp*/ )
+        external
+        pure
+        returns (uint[] memory)
+    {
+        return new uint[](0);
     }
 
-    function getOperatorCount(
-        OperatorSet memory, /*operatorSet*/
-        uint32 /*referenceTimestamp*/
-    ) external pure returns (uint256) {
+    function getOperatorCount(OperatorSet memory, /*operatorSet*/ uint32 /*referenceTimestamp*/ ) external pure returns (uint) {
         return 0;
     }
 }
